@@ -1,5 +1,5 @@
 module contactForm
 
-go 1.19
+go 1.20
 
 require gopkg.in/ezzarghili/recaptcha-go.v4 v4.3.0
