@@ -2,8 +2,6 @@ package contactForm
 
 import (
 	"encoding/json"
-	"io"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"setup"
@@ -72,8 +70,23 @@ func SaveContact(c *gin.Context) {
 
 // saveContact writes the a ContactForm to a json file of ContactForms
 func saveContact(contact ContactForm) {
-	// get existing contacts
-	contactList := loadContacts()
+	// file path
+	prefix := os.Getenv("LOG_FILE_PATH")
+	filepath := prefix + os.Getenv("CONTACT_FORM_FILE_NAME")
+
+	// Open our jsonFile
+	jsonFile, err := os.ReadFile(filepath)
+	if err != nil {
+		setup.LogCommon(err).Error("Open contact form file")
+	}
+
+	var contactList ContactForms
+	// map from bytes to struct
+	err = json.Unmarshal(jsonFile, &contactList)
+	if err != nil {
+		setup.LogCommon(err).Error("Unmarshall")
+	}
+
 	// add new contact
 	contactList.ContactForms = append(contactList.ContactForms, contact)
 
@@ -84,37 +97,10 @@ func saveContact(contact ContactForm) {
 	}
 
 	// save to file
-	err = ioutil.WriteFile("./internal/contactForm/contacts.json", contactsJSON, 0644)
+	err = os.WriteFile(filepath, contactsJSON, 0664)
 	if err != nil {
 		setup.LogCommon(err).Error("Write File")
 	} else {
 		setup.LogCommon(nil).Info("Saved contact form")
 	}
-}
-
-// loadContacts reads a JSON file of contacts to a ContactForms
-func loadContacts() ContactForms {
-	// Open our jsonFile
-	jsonFile, err := os.Open("./internal/contactForm/contacts.json")
-	// if we os.Open returns an error then handle it
-	if err != nil {
-		setup.LogCommon(err).Error("Open File")
-	}
-	// defer the closing of our jsonFile so that we can parse it later on
-	defer jsonFile.Close()
-
-	// read our opened jsonFile as a byte array.
-	byteValue, err := io.ReadAll(jsonFile)
-	if err != nil {
-		setup.LogCommon(err).Error("Read File")
-	}
-
-	var contacts ContactForms
-	// map from bytes to struct
-	err = json.Unmarshal(byteValue, &contacts)
-	if err != nil {
-		setup.LogCommon(err).Error("Unmarshall")
-	}
-
-	return contacts
 }
