@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
-import shoppingCartReducer from "./components/ShoppingCart/shoppingCartReducer";
+import shoppingCartReducer from "./shoppingCart";
 
 // The only thing you should need to do in this file is add reducers
 // See - https://redux-toolkit.js.org/tutorials/typescript
